requirements.txt
Django>=4.0
djangorestframework>=3.14.0
requests>=2.31.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
dnspython>=2.4.2
python-whois>=0.8
dnspython>=2.4.2
beautifulsoup4>=4.12.2
python-whois>=0.8


